+++ EnergyVisualizer/Source/PluginProcessor.h
@@ -75,6 +75,10 @@ public:
 
     const float getPeakLevelSetting() { return *peakLevel; }
     const float getDynamicRange() { return *dynamicRange; }
+    const float getHoldRMSSetting(){
+        if (*holdMax >= 0.5f) return true;
+        else return false;
+    }
 
     std::vector<float> rms;
     juce::Atomic<juce::Time> lastEditorTime;
@@ -86,6 +90,8 @@ private:
     std::atomic<float>* useSN3D;
     std::atomic<float>* peakLevel;
     std::atomic<float>* dynamicRange;
+    std::atomic<float>* holdMax;
+    std::atomic<float>* RMStimeConstant;
 
     float timeConstant;